fix(crypto): restore throw contract in prices(in:) + fix schema version test
Three test failures on this branch:

1. ProfileIndexSchemaV3Tests.versionIsLatest: updated expected version from 5
   to 6 to match the `v6_purge_rate_caches` migration added on this branch.

2 & 3. CryptoPriceServiceAttributionTests rangePrices*: `coverRangeContiguously`
   broke the attributed-error contract introduced in `fetchRange`. Added
   `lastFetchError` capture across both endpoint loops; after both loops,
   if either endpoint is still uncovered (checked via ContiguousFetchPlanner),
   the captured error is rethrown — mirroring `resolveAfterExtension`'s
   pattern so `prices(in:)` surfaces the same attributed WalletSyncError as
   `price(on:)`.

   Side-effect of the attribution fix: `CryptoPriceService+FetchRange.swift`
   exceeded the 400-line SwiftLint limit. Extracted the background-warming
   methods (`warmRange`, `warmStep`, `fetchSubRangeWarming`) into a new
   `CryptoPriceService+Warming.swift` file to bring both files under 400 lines.
   `parseInterval` was widened from `private` to `internal` so the new file
   can call it.

Full macOS suite: 4188 tests passed, 0 failed.
